This game simulates a Quincunx game.
I had a project asked me to make a Quincunx game.  This is a game where you have a board with several nails put into it that form a pyramid.  At the bottom of the board there are slots that the balls place into.  The player then takes balls and drops them at the top of the board so that the balls bounce on the nails.  Each time the ball hits a pin there is a 50/50 chance the ball with fall left (L) or right (R).  At the end the balls drop into slots.  
This game asks for the amount of balls used, then for the number of slots at the bottom of the quincunx board.  After this the program simulates the game and shows the direction each ball falls by indicating “L” for left or “R” for Right.  At the very bottom of the display I have simulated text that shows the slots and how they are stacked.