A small game I made.
This game was made using the JMonkey engine.  The project was to make a video game level.  I was paired with one partner, Tyler Cheatham.  He and I had made the level using code that was found in the tutorial of the JMonkey engine.  He fixed up some of the looks, and the file saving option.  I made all the objects, the HUD, and coded how all the objects behave.