Calendar App
This was the final project of one of my classes.  Here I had to make a calendar app that took the date from the computer, and used that to make a calendar.  The project focused on just displaying a calendar correctly.  I went a little further though, and added the options to go back or forward month by month.  I also added a feature where the current date has a red border around it.