Sorcerers Cave
I really did not like this project, but only because of the layout.  This program takes formatted .txt files as input and extracts data from them to build a mock strategy game that is similar to "Dungeons and Dragons."  It makes several objects, and the part I don't like is that I had to extend them in very inefficient ways.  At the end I had to write a part the managed jobs performed by members, and to do this I explored the use of multiprocessor programming.